Main: PixelUtil::getMemorySize - multiply as size_t
to avoid overlflows

+++ OgreMain/src/OgrePixelFormat.cpp
@@ -185,7 +185,7 @@ namespace Ogre {
         }
         else
         {
-            return width*height*depth*getNumElemBytes(format);
+            return size_t(width)*height*depth*getNumElemBytes(format);
         }
     }
     //-----------------------------------------------------------------------
